+++ bin/core/routes/main/mainRoute.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import main from '../../controller/mainController.js';
+
+const mainRouter = express.Router();
+
+mainRouter.get('/', main.view);
+
+export default mainRouter;
