+++ bin/core/views/board.ejs
