+++ bin/core/views/category.ejs
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1.0, font-src 'self'"
+    />
+    <title>카테고리</title>
+  </head>
+</html>
