+++ bin/core/views/liveChat/form/createroomform.ejs
@@ -1,0 +1,4 @@
+<form>
+  <input />
+  <button></button>
+</form>
