+++ bin/core/views/liveChat/form/registerform.ejs
@@ -1,0 +1,14 @@
+<form id="add_user" class="add_user">
+  <header><h3>사용자 추가</h3></header>
+  <div>
+    <label for="input_username">사용자 이름 입력</label>
+    <input
+      type="text"
+      id="input_username"
+      class="input_username"
+      placeholder="사용자명을 입력하세요"
+      required
+    />
+  </div>
+  <button type="submit">사용자 추가</button>
+</form>
