+++ bin/core/views/liveChat/header/livechatCategory.ejs
@@ -1,0 +1,4 @@
+<header>
+  <h1>라이브챗 카테고리</h1>
+  <%- include('../form/registerform')%>
+</header>
