+++ bin/core/views/liveChat/list/roomlist.ejs
@@ -1,0 +1,4 @@
+<div>
+  <header><h2>방 리스트</h2></header>
+  <ul class="room_list_guide"></ul>
+</div>
