+++ bin/core/views/liveChat/list/userlist.ejs
@@ -1,0 +1,4 @@
+<div>
+  <header><h2>유저리스트</h2></header>
+  <ul class="user_list_guide"></ul>
+</div>
