+++ bin/core/views/liveChat/liveChatRoom.ejs
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Live Chat Main</title>
+  </head>
+  <body>
+    <script src="/socket.io/socket.io.js"></script>
+    <script src="/static/public/liveChat/js/liveSourcePack.js"></script>
+  </body>
+</html>
