+++ bin/core/views/posts/post_editpage.ejs
