+++ bin/core/views/posts/post_pagination.ejs
