+++ bin/core/views/posts/post_writepage.ejs
