--- static/public/liveinvader/js/liveSourcePack.js
@@ -1,7 +1,0 @@
-const socket = io();
-
-const register_user_form = document.querySelector('.register-form');
-const register_username_input = document.querySelector('.username-input');
-const create_room_btn = document.querySelector('.create_chat_room');
-const create_room_form = document.querySelector('.create_room');
-const active_room_list = document.querySelector('.active_room_list');
