+++ utils/[Class]livechatData/liveRoom.js
@@ -1,0 +1,3 @@
+class Room {
+  constructor() {}
+}
