--- views/category.ejs
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, font-src 'self'"
-    />
-    <title>카테고리</title>
-  </head>
-  <body>
-    <nav>
-      <ul>
-        <li>메뉴1</li>
-        <li>메뉴2</li>
-        <li>메뉴3</li>
-        <li>메뉴4</li>
-        <li>메뉴5</li>
-      </ul>
-    </nav>
-  </body>
-</html>
