--- views/liveinvader/liveChat.ejs
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<head>
-  <meta charset="utf-8" />
-  <title>Live Invader에서 재미를 누려보자</title>
-</head>
-<body>
-  <h1><%= title %></h1>
-  <aside id="user-area"><%- include('./liveChatUserList') %></aside>
-  <section id="msg-area">
-    <div>
-      <ul class="text-box"></ul>
-    </div>
-  </section>
-  <section id="chatform-area">
-    <form action="" class="chatform-box">
-      <input
-        id="msg"
-        class="input-text"
-        autocomplete="off"
-        placeholder="메세지를 입력하세요"
-      />
-      <button class="message_btn" type="submit">입력</button>
-    </form>
-  </section>
-  <!-- socket.io 를 Global영역으로 가져와 사용 -->
-  <script src="/socket.io/socket.io.js"></script>
-  <!-- socket.io 를 활용한 코드 -->
-  <script
-    src="/static/public/liveinvader/js/liveInvader.js"
-    type="module"
-  ></script>
-</body>
