--- views/liveinvader/liveChatHead.ejs
@@ -1,1 +1,0 @@
-<header></header>
