--- views/liveinvader/liveChatLoginUser.ejs
@@ -1,8 +1,0 @@
-<div>
-  <header>
-    <h1>메인 참여자</h1>
-  </header>
-  <div>
-    <ul class="joined-mainpage-user-list"></ul>
-  </div>
-</div>
