--- views/liveinvader/liveChatRoomList.ejs
@@ -1,8 +1,0 @@
-<div>
-  <header>
-    <h1>방 목록</h1>
-  </header>
-  <div>
-    <ul class="active_room_list"></ul>
-  </div>
-</div>
