--- views/liveinvader/liveChatUserList.ejs
@@ -1,8 +1,0 @@
-<div>
-  <header>
-    <h1>참여자</h1>
-  </header>
-  <div>
-    <ul class="joined-user-list"></ul>
-  </div>
-</div>
