--- views/liveinvader/liveCreateRoomForm.ejs
@@ -1,21 +1,0 @@
-<section id="register-container">
-  <header>
-    <h1>채팅방 생성</h1>
-  </header>
-  <form action="<%= %>" class="create_room">
-    <article>
-      <label for="chat_room_subject_input">채팅방 이름</label>
-      <input
-        type="text"
-        class="chat_roomname_input"
-        id="chat_room_subject_input"
-        placeholder="생성할 방 이름 입력"
-        maxlength="30"
-        required
-      />
-    </article>
-    <div>
-      <button type="button" class="create_chat_room">방 생성</button>
-    </div>
-  </form>
-</section>
