--- views/liveinvader/liveMain.ejs
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head> </head>
-  <body>
-    <header>
-      <h1>라이브 인베이더</h1>
-    </header>
-    <aside>
-      <%- include('./liveChatLoginUser') %> <%- include('./liveChatRoomList') %>
-    </aside>
-    <div><%- include('./liveRegisterForm') %></div>
-    <div><%- include('./liveCreateRoomForm') %></div>
-    <!-- socket.io 를 Global영역으로 가져와 사용 -->
-    <script src="/socket.io/socket.io.js"></script>
-    <!-- socket.io 를 활용한 코드 -->
-    <script
-      src="/static/public/liveinvader/js/warehouse.js"
-      type="module"
-    ></script>
-  </body>
-</html>
