--- views/liveinvader/liveRegisterForm.ejs
@@ -1,17 +1,0 @@
-<section id="register-container">
-  <header>
-    <h1>가입하기</h1>
-  </header>
-  <form class="register-form">
-    <label for="username-input">사용자 명</label>
-    <input
-      type="text"
-      name="username"
-      class="username-input"
-      id="username-input"
-      placeholder="사용자명 입력"
-      maxlength="18"
-    />
-    <button type="submit" class="username-submit">사용자 생성</button>
-  </form>
-</section>
