--- views/main.ejs
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="author" content="HP Laptop" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, font-src 'self'"
-    />
-    <title>HP 커뮤니티</title>
-  </head>
-  <body>
-    <div><%= title %></div>
-    <%- include('./category') %>
-    <a href="/boards/list/1">갤러리</a>
-    <a href="/liveinvader/">라이브 인베이더</a>
-  </body>
-</html>
